+++ src/csharp/Microsoft.Spark/ML/Feature/Base.cs
@@ -4,8 +4,6 @@
 
 using System.Linq;
 using System.Reflection;
-using Microsoft.Spark.Sql;
-using Microsoft.Spark.Sql.Types;
 using Microsoft.Spark.Interop;
 using Microsoft.Spark.Interop.Ipc;
 
@@ -123,195 +121,10 @@ namespace Microsoft.Spark.ML.Feature
     }
 
     /// <summary>
-    /// <see cref="ScalaPipelineStage"/> A stage in a pipeline, either an Estimator or a Transformer.
-    /// </summary>
-    public abstract class ScalaPipelineStage : Params
-    {
-        internal ScalaPipelineStage(string className) : base(className)
-        {
-        }
-
-        internal ScalaPipelineStage(string className, string uid) : base(className, uid)
-        {
-        }
-
-        internal ScalaPipelineStage(JvmObjectReference jvmObject) : base(jvmObject)
-        {
-        }
-
-        /// <summary>
-        /// Check transform validity and derive the output schema from the input schema.
-        ///
-        /// We check validity for interactions between parameters during transformSchema
-        /// and raise an exception if any parameter value is invalid.
-        ///
-        /// Typical implementation should first conduct verification on schema change and
-        /// parameter validity, including complex parameter interaction checks.
-        /// </summary>
-        /// <param name="schema">
-        /// The <see cref="StructType"/> of the <see cref="DataFrame"/> which will be transformed.
-        /// </param>
-        /// <returns>
-        /// The <see cref="StructType"/> of the output schema that would have been derived from the
-        /// input schema, if Transform had been called.
-        /// </returns>
-        public virtual StructType TransformSchema(StructType schema) =>
-             new StructType(
-                (JvmObjectReference)Reference.Invoke(
-                    "transformSchema",
-                    DataType.FromJson(Reference.Jvm, schema.Json)));
-    }
-
-    /// <summary>
-    /// <see cref="ScalaTransformer"/> Abstract class for transformers that transform one dataset into another.
-    /// </summary>
-    public abstract class ScalaTransformer : ScalaPipelineStage
-    {
-        internal ScalaTransformer(string className) : base(className)
-        {
-        }
-
-        internal ScalaTransformer(string className, string uid) : base(className, uid)
-        {
-        }
-
-        internal ScalaTransformer(JvmObjectReference jvmObject) : base(jvmObject)
-        {
-        }
-
-        /// <summary>
-        /// Executes the transformer and transforms the DataFrame to include new columns.
-        /// </summary>
-        /// <param name="dataset">The Dataframe to be transformed.</param>
-        /// <returns>
-        /// <see cref="DataFrame"/> containing the original data and new columns.
-        /// </returns>
-        public virtual DataFrame Transform(DataFrame dataset) =>
-            new DataFrame((JvmObjectReference)Reference.Invoke("transform", dataset));
-    }
-
-    /// <summary>
-    /// A helper interface for ScalaEstimator, so that when we have an array of ScalaEstimators
-    /// with different type params, we can hold all of them with Estimator&lt;object&gt;.
-    /// </summary>
-    public interface Estimator<out M>
-    {
-        M Fit(DataFrame dataset);
-    }
-
-    /// <summary>
-    /// Abstract Class for estimators that fit models to data.
-    /// </summary>
-    /// <typeparam name="M"/>
-    public abstract class ScalaEstimator<M> : ScalaPipelineStage, Estimator<M> where M : ScalaModel<M>
-    {
-        internal ScalaEstimator(string className) : base(className)
-        {
-        }
-
-        internal ScalaEstimator(string className, string uid) : base(className, uid)
-        {
-        }
-
-        internal ScalaEstimator(JvmObjectReference jvmObject) : base(jvmObject)
-        {
-        }
-
-        /// <summary>
-        /// Fits a model to the input data.
-        /// </summary>
-        /// <param name="dataset">input dataset.</param>
-        /// <returns>fitted model</returns>
-        public virtual M Fit(DataFrame dataset) =>
-            WrapAsType<M>((JvmObjectReference)Reference.Invoke("fit", dataset));
-    }
-
-    /// <summary>
-    /// A helper interface for ScalaModel, so that when we have an array of ScalaModels
-    /// with different type params, we can hold all of them with Model&lt;object&gt;.
-    /// </summary>
-    public interface Model<out M>
-    {
-        bool HasParent();
-    }
-
-    /// <summary>
-    /// A fitted model, i.e., a Transformer produced by an Estimator.
-    /// </summary>
-    /// <typeparam name="M">
-    /// Model Type.
-    /// </typeparam>
-    public abstract class ScalaModel<M> : ScalaTransformer, Model<M> where M : ScalaModel<M>
-    {
-        internal ScalaModel(string className) : base(className)
-        {
-        }
-
-        internal ScalaModel(string className, string uid) : base(className, uid)
-        {
-        }
-
-        internal ScalaModel(JvmObjectReference jvmObject) : base(jvmObject)
-        {
-        }
-
-        /// <summary>
-        /// Sets the parent of this model.
-        /// </summary>
-        /// <param name="parent">The parent of the ScalaModel to be set</param>
-        /// <returns>type parameter M</returns>
-        public M SetParent(ScalaEstimator<M> parent) =>
-            WrapAsType<M>((JvmObjectReference)Reference.Invoke("setParent", parent));
-
-        /// <summary>
-        /// Indicates whether this Model has a corresponding parent.
-        /// </summary>
-        /// <returns>bool</returns>
-        public bool HasParent() =>
-            (bool)Reference.Invoke("hasParent");
-    }
-
-    /// <summary>
-    /// <see cref="ScalaEvaluator"/> Abstract Class for evaluators that compute metrics from predictions.
-    /// </summary>
-    public abstract class ScalaEvaluator : Params
-    {
-        internal ScalaEvaluator(string className) : base(className)
-        {
-        }
-
-        internal ScalaEvaluator(string className, string uid) : base(className, uid)
-        {
-        }
-
-        internal ScalaEvaluator(JvmObjectReference jvmObject) : base(jvmObject)
-        {
-        }
-
-        /// <summary>
-        /// Evaluates model output and returns a scalar metric.
-        /// The value of isLargerBetter specifies whether larger values are better.
-        /// </summary>
-        /// <param name="dataset">a dataset that contains labels/observations and predictions.</param>
-        /// <returns>metric</returns>
-        public virtual double Evaluate(DataFrame dataset) =>
-            (double)Reference.Invoke("evaluate", dataset);
-
-        /// <summary>
-        /// Indicates whether the metric returned by evaluate should be maximized 
-        /// (true, default) or minimized (false).
-        /// A given evaluator may support multiple metrics which may be maximized or minimized.
-        /// </summary>
-        /// <returns>bool</returns>
-        public bool IsLargerBetter =>
-            (bool)Reference.Invoke("isLargerBetter");
-    }
-
-    /// <summary>
-    /// DotnetHelper is used to hold basic general helper functions that
+    /// DotnetUtils is used to hold basic general helper functions that
     /// are used within ML scope.
     /// </summary>
-    public class DotnetHelper
+    public class DotnetUtils
     {
         /// <summary>
         /// Helper function for getting the exact class name from jvm object.
