+++ src/csharp/Microsoft.Spark/ML/Param/Param.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
 using Microsoft.Spark.Interop;
 using Microsoft.Spark.Interop.Ipc;
 
@@ -74,42 +73,5 @@ namespace Microsoft.Spark.ML.Feature.Param
         /// </summary>
         /// <returns>The UID of the parent oject that this <see cref="Param"/> belongs to</returns>
         public string Parent => (string)Reference.Invoke("parent");
-    }
-
-    /// <summary>
-    /// A param to value map.
-    /// </summary>
-    public class ParamMap : IJvmObjectReferenceProvider
-    {
-        private static readonly string s_ParamMapClassName = "org.apache.spark.ml.param.ParamMap";
-
-        /// <summary>
-        /// Creates a new instance of a <see cref="ParamMap"/>
-        /// </summary>
-        public ParamMap() : this(SparkEnvironment.JvmBridge.CallConstructor(s_ParamMapClassName))
-        {
-        }
-
-        public ParamMap(JvmObjectReference jvmObject) => Reference = jvmObject;
-
-        public JvmObjectReference Reference { get; private set; }
-
-        /// <summary>
-        /// Puts a (param, value) pair (overwrites if the input param exists).
-        /// </summary>
-        /// <param name="param">The param to be add</param>
-        /// <param name="value">The param value to be add</param>
-        public ParamMap Put<T>(Param param, T value) =>
-            WrapAsParamMap((JvmObjectReference)Reference.Invoke("put", param, value));
-
-        /// <summary>
-        /// Returns the string representation of this ParamMap.
-        /// </summary>
-        /// <returns>representation as string value.</returns>
-        public override string ToString() =>
-            (string)Reference.Invoke("toString");
-
-        private static ParamMap WrapAsParamMap(object obj) =>
-            new ParamMap((JvmObjectReference)obj);
     }
 }
