+++ src/scala/microsoft-spark-3-0/src/main/scala/org/apache/spark/api/dotnet/DotnetUtils.scala
@@ -9,19 +9,10 @@ package org.apache.spark.api.dotnet
 import org.apache.spark.ml._
 import scala.collection.JavaConverters._
 
-/** DotnetHelper object that hosts some helper functions
+/** DotnetUtils object that hosts some helper functions
   * help data type conversions between dotnet and scala
   */
-object DotnetHelper {
-
-  /** A helper function to let pipeline accept java.util.ArrayList
-    * format stages in scala code
-    * @param pipeline - The pipeline to be set stages
-    * @param value - A java.util.ArrayList of PipelineStages to be set as stages
-    * @return The pipeline
-    */
-  def setPipelineStages(pipeline: Pipeline, value: java.util.ArrayList[_ <: PipelineStage]): Pipeline =
-    pipeline.setStages(value.asScala.toArray)
+object DotnetUtils {
 
   /** A helper function to convert scala Map to java.util.Map
     * @param value - scala Map
